Ya se puede crear y leer
Ya se crea correctamente el fichero xml y se lee, ahora falta cargar los
datos a la app

+++ src/Log/EscribirXml.java
@@ -7,24 +7,35 @@ import org.jdom2.output.*;
  * @author Lyderys
  */
 public class EscribirXml {
-    public static void Escribir(){
+    String sFichero = "config.xml";
+    File fichero = new File(sFichero);
+    public void Escribir(){
         Element config = new Element("config");
         Document doc = new Document(config);
-        doc.setRootElement(config);
+        //doc.setRootElement(config);
         Element configuracion = new Element("configuracion");
-        configuracion.addContent(new Element("puertom").setText("PUERTO_DE_MENSAJERIA"));
-        configuracion.addContent(new Element("puertop").setText("PUERTO_DE_CONTROL"));
-        configuracion.addContent(new Element("numuser").setText("NUMERO_DE_USUARIOS"));
-        configuracion.addContent(new Element("tambuffer").setText("TAMAÑO_DE_BUFFER"));
-        configuracion.addContent(new Element("log").setText("FICHERO_PARA_EL_LOG"));
-        configuracion.addContent(new Element("infocon").setText("INFORMACION_DE_CONEXION"));
+        Element campos = new Element("campos");
+        configuracion.addContent(campos);
+        campos.addContent(new Element("puertom").setText("PUERTO_DE_MENSAJERIA"));
+        campos.addContent(new Element("puertoc").setText("PUERTO_DE_CONTROL"));
+        campos.addContent(new Element("numuser").setText("NUMERO_DE_USUARIOS"));
+        campos.addContent(new Element("tambuffer").setText("TAMAÑO_DE_BUFFER"));
+        campos.addContent(new Element("log").setText("FICHERO_PARA_EL_LOG"));
+        campos.addContent(new Element("infocon").setText("INFORMACION_DE_CONEXION"));
         doc.getRootElement().addContent(configuracion);
         XMLOutputter xmlOutput = new XMLOutputter();
         xmlOutput.setFormat(Format.getPrettyFormat());
-        try {
+        if(fichero.exists()){
+            
+            //codigo para arrancar la carga de datos
+        }else{
+            
+           try {
             xmlOutput.output(doc, new FileWriter("config.xml"));
-        } catch (IOException ex) {
+            } catch (IOException ex) {
             System.out.println(ex.getMessage());
+            } 
         }
+        
     }
 }
